Simplify switching the astro.io backend
Switching this has always been possible, but this PR makes it simpler
by (a) adding a dummy_backend (similar to what we already have for plotting)
to ensure that sherpa.astro.io can always be imported even if no fits reader
is present, and (b) adding docs and tests on how to switch.

Since switching is relatively easy, there is no need for a configuration item any longer.
Instead, I just hardcode an order of imports to try (first crates, then pyfits, thnen dummy)
and if the user does not like it, they can switch after. Both backends are
functionally identical anyway (except for rare cases where some fits file
triggers a bug in one, but not the other backend and we did not have time
to fix that upstream yet.)
Thus, I removed the configuration item from the sherpa.rc files.