Always import backend as "io.backend"
That way, if I assign a new backend in "sherpa.astro.io.backend = mybackend"
than that change will be picked up. Otherwise, that was set when this module
is imported will be used.
Maybe not that important for a test case (since they get executed pretty much
right afte rbeing imported) but that's the way it should be.